feat: add eth_maxPriorityFeePerGas to method whitelist
This change adds the `eth_maxPriorityFeePerGas` JSON-RPC method to the
list of allowed methods. This is a standard EIP-1559 method that allows
clients to get a suggestion for the maximum priority fee.

By whitelisting this method, clients that support EIP-1559 can now
interact with the node to get recommended gas prices, enabling more
efficient and predictable transaction fees.